+++ src/lib.rs
@@ -43,9 +43,9 @@ mod serde;
 mod util;
 
 pub use client::customers::{
-    Address, AddressRequest, CreateCustomerRequest, Customer, CustomerId,
-    CustomerPaymentProviderRequest, LedgerEntryRequest, LedgerEntry,
-    PaymentProvider, UpdateCustomerRequest, VoidReason,
+    AddIncrementCreditLedgerEntryRequestParams, AddVoidCreditLedgerEntryRequestParams, Address,
+    AddressRequest, CreateCustomerRequest, Customer, CustomerId, CustomerPaymentProviderRequest,
+    LedgerEntry, LedgerEntryRequest, PaymentProvider, UpdateCustomerRequest, VoidReason,
 };
 pub use client::events::{
     AmendEventRequest, Event, EventPropertyValue, EventSearchParams, IngestEventDebugResponse,